fixing issues: MessageRepositoryImpl.kt

+++ client/feature/chat/src/main/java/com/whatsappclone/feature/chat/data/MessageRepositoryImpl.kt
@@ -78,7 +78,7 @@ class MessageRepositoryImpl @Inject constructor(
         mediaThumbnailUrl: String?,
         mediaMimeType: String?,
         mediaSize: Long?,
-        mediaDuration: Long?
+        mediaDuration: Int?
     ): AppResult<Unit> {
         val currentUserId = getCurrentUserId() ?: return AppResult.Error(
             code = ErrorCode.UNAUTHORIZED,
